hh/requirements.txt
asgiref==3.3.4
channels==3.0.4
channels-redis==3.3.1
Django==3.2.10
django-cities-light==3.9.0
django-crispy-forms==1.13.0
django-el-pagination==3.3.0
django-floppyforms==1.9.0
django-notifications-hq==1.6.0
django-smart-selects==1.5.9
environs==9.3.5
Faker==11.3.0
marshmallow==3.14.1
mixer==7.2.0
Pillow==9.0.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
feedparser
django-widget-tweaks==1.4.12
psycopg2-binary==2.9.1
